OeX_ES-33: Refactoring due removing custom response class to separete module.
