+++ course_discovery/apps/edx_haystack_extensions/distinct_counts/query.py
@@ -1,7 +1,8 @@
 from django.conf import settings
 from elasticsearch_dsl.connections import get_connection
 
-from course_discovery.apps.edx_haystack_extensions.viewsets import DistinctDSLResponse, FacetedSearch
+from course_discovery.apps.edx_haystack_extensions.response import DistinctDSLResponse
+from course_discovery.apps.edx_haystack_extensions.search import FacetedSearch
 
 
 class DistinctCountsSearchQuerySet(FacetedSearch):
